Rewrite both paths into comprehensions
After taking a quick peek at the answer, I changed the code to create a
function that returns the transformed versions of either size or
color(s) of the fruit or vegetable.

It feels a bit more challening to adapt the algorithm with
comprehensions in mind, so I'll figure out how to do this.