Do steps 1-3 of algo for problem 11
Starting out with using `for` loops, I first created an empty list and
assigned it to the `all_vowels` variable. This will be the final list
that will be returned at the end of the function.

Then I created a string of `'aeiou'` and assigned it to a variable
called `vowels`.

Then I created the outer `for` loop to target the values of the input
dictionary using the `values` dictionary method.

Then I tested using `print` - which returned:
```
['the', 'quick']
['brown', 'fox']
['jumped']
['over', 'the', 'lazy', 'dog']
```
Which indicates that it works.