Continue with steps of problem 11
I created a nested `for` loop to target the strings in the nested lists
of the input dictionary. Then I used the `print` function to test what
`str` returns.

It returns:
```
the
quick
brown
fox
jumped
over
the
lazy
dog
```
Which are all of the strings contained in the input dictionary. So it
works.