Find bug and fix
The issue: the beginning message at the start of `play_match` was not
displaying; It was going straight to the playing the game.

The issue was where `os.system('clear')` as placed. It was at the
beginning of the 'display_board' function, which I think may have made
it so that the beginning message was displayed but the terminal was
cleared too quickly to notice.

Moved it to `play_round` function.

Now halfway there to figuring out how to get to where I want it to be.