Fix `find_at_risk_square` function
After some testing and realizing that that the function I had only
checks to see if whatever square is assigned to `sq1` & `sq2` are marked
with `X`, with the last square, `sq3`, being blank. It doesn't check for
other combinations.

After looking at LS's solution for the function, I realized that the
approach I had was not adequate for the job. I had to come up with a
way to check if ANY combination of each sub-list were occupied by two
`X`s.

The `find_at_risk_square` function in the solution:

Takes two parameters, `line` (each sub-list in `WINNING_LINES`) and
`board` (the current state of the board).

Then it creates a new list out of the marker version of each square
number/element of `line`. This means that it will convert, say, the
list `[1, 2, 3]` in its current marked form, say `[X, O, ' ']`. Then
this is assigned to `markers_in_line`.

Then, if there are a total of two `'X'`s in `markers_in_line`, each
marker in the list will be checked for `INITIAL_MARKER`. If it is found
(meaning it’s not occupied by an `X` or an `O`), then the square
number/element of that sub-list in the current state of the board is
returned.

Otherwise, if `markers_in_line` are not occupied by two `X`s, then the
function will return `None`.