Modify `computer_chooses_square` function
Put in LS's solution. The function works this way:
The `computer_chooses_square` function takes one parameter, `board`,
which is the current state of the board.

If the number of available squares is none, as indicated by the list
returned by `empty_squares(board)`, then the function just returns/ends
as the computer cannot choose anything.

Then `None`, a falsy value, is assigned to the variable `square`.

Each sub-list (`line`) in `WINNING_LINES` is iterated over. The
`find_at_risk_square` function is called on each sub-list, with the
sub-list & current state of the board being passed as arguments. The
returning value (either a square number/element or `None`) is then
assigned to `square`.

If `square` is a truthy value, meaning two `X`s were found & an empty
square was detected, then the loop breaks and goes onto the next part
of the function.

Then the `if` statement on line 11 has the computer go ahead and pick a
random square if `square` is `None`, meaning there is no immediate
threat (the `not` operator here flips that `None` to a truthy value,
`True`, causing the code in the body to execute).

But if `square` is assigned to a truthy value (a square number), then
that value is used to replace the square there with the computer’s
marker, `O`.