Work on "Computer AI: Offense" bonus feature
Also moved the input to go onto the next round and clearing the board to
only trigger if neither player nor computer has reached 5 points because
it was showing up even after one of them did.

Started prepping for the Offense bonus feature. Now `find_at_risk_
square` returns the square number that is empty if there are either
two Xs or two Os in the sublist of the `WINNING_LINES`.