Incorporate LS's Solution for computer offense
The change to `computer_chooses_square` function works because now,
the variable `square` will start out with the value `None`, then it
goes onto see if playing defense or offense will return a truthy value.
Otherwise, the computer will mark a random square.

To explain line by line: the defense is played first, to block any
immediate threats. For each sublist in `WINNING_LINES`, the `find_at_
risk` function will be called, with the sublist, the current state of
the board, and `'X'` passed as the arguments. The returning value is
reassigned to `square`. If `square` refers to a truthy value, a square
number that is empty, then the loop is broken, and that square number is
used to mark that square with a `'O'`.

After every single sublist in `WINNING_LINES` is checked for threats,
and there wasn't any immediate threats, now `square` is a falsy value
and the body of the `if` statement is executed when the `not` operator
is used to flip it to a truthy value.

Then the sub-lists are checked once again to see if any of them contain
two `'O'`s. If there are any, then `square` is reassigned with the
returning value, the square number. The code goes onto use that value to
mark that square with an `'O'`.

If niether of the defense nor the offense code returns a truthy value,
then `square` is still `None`, and a random square is picked and that
square number/element is reassigned to `square`. This is used to mark
that square with the computer's marker.