fix: convert to spaces & rewrite function
Realized that the `face_card_names` function was working with the
original list passed as the argument (which we need later to perform
calculations in the `total` function) instead of a new list created just
for displaying the hand in human-readable format.

So changed it to create a new list instead.