refactor: change `busted` to accept new variables
The second part of the bonus features assignment "Minimizing Calls to
Total" calls for the modification of the `busted` and `detect_result`
(or `determine_winner` in my case) functions to accept the newly created
local variables instead of calls to the `total` function.

This commit was for the `busted` function; However, since my `detect_
result` function was already written to use total points instead of
a player's hand, it will not be modified.